chore: extract file selector logic into a composable
